+++ src/kernel/LinearCharKernel.cpp
@@ -114,7 +114,7 @@ bool CLinearCharKernel::init_optimization(INT num_suppvec, INT* sv_idx, REAL* al
 		assert(avec);
 
 		for (int j=0; j<num_feat; j++)
-			normal[j]+=alphas[i]*avec[j];
+			normal[j]+= alphas[i] * ((double) avec[j]);
 
 		((CCharFeatures*) lhs)->free_feature_vector(avec, 0, afree);
 	}
@@ -140,7 +140,7 @@ REAL CLinearCharKernel::compute_optimized(INT idx_b)
 	double result=0;
 	{
 		for (INT i=0; i<blen; i++)
-			result+= ((LONG) normal[i]) * ((LONG) bvec[i]);
+			result+= normal[i] * ((double) bvec[i]);
 	}
 	result/=scale;
